+++ database/factories/UserFactory.php
@@ -1,0 +1,12 @@
+<?php
+
+use Chriscreates\Projects\Models\User;
+use Faker\Generator as Faker;
+
+$factory->define(User::class, function (Faker $faker) {
+    return [
+        'name' => $faker->name,
+        'email' => $faker->email,
+        'password' => $faker->password,
+    ];
+});