+++ tests/ProjectTest.php
@@ -5,14 +5,16 @@ namespace Chriscreates\Projects\Tests;
 use Chriscreates\Projects\Models\Project;
 use Chriscreates\Projects\Models\Status;
 use Chriscreates\Projects\Models\User;
-use Illuminate\Support\Facades\Hash;
+use Illuminate\Foundation\Testing\RefreshDatabase;
 
 class ProjectTest extends TestCase
 {
+    use RefreshDatabase;
+
     /** @test */
     public function it_has_a_title()
     {
-        $project = Project::create(['title' => 'Title']);
+        $project = factory(Project::class)->create(['title' => 'Title']);
 
         $this->assertEquals($project->title, 'Title');
     }
@@ -20,10 +22,8 @@ class ProjectTest extends TestCase
     /** @test */
     public function it_has_a_description()
     {
-        $project = Project::create([
-            'title' => 'Title',
-            'description' => 'This is a description',
-        ]);
+        $project = factory(Project::class)
+        ->create(['description' => 'This is a description']);
 
         $this->assertEquals($project->description, 'This is a description');
     }
@@ -31,10 +31,8 @@ class ProjectTest extends TestCase
     /** @test */
     public function it_has_a_notes()
     {
-        $project = Project::create([
-            'title' => 'Title',
-            'notes' => 'These are some notes',
-        ]);
+        $project = factory(Project::class)
+        ->create(['notes' => 'These are some notes']);
 
         $this->assertEquals($project->notes, 'These are some notes');
     }
@@ -42,10 +40,8 @@ class ProjectTest extends TestCase
     /** @test */
     public function it_has_togglable_visiblitity()
     {
-        $project = Project::create([
-            'title' => 'Title',
-            'visible' => true,
-        ]);
+        $project = factory(Project::class)
+        ->create(['visible' => true]);
 
         $this->assertSame($project->visible, true);
     }
@@ -53,10 +49,8 @@ class ProjectTest extends TestCase
     /** @test */
     public function it_has_a_started_at_date()
     {
-        $project = Project::create([
-            'title' => 'Title',
-            'started_at' => now(),
-        ]);
+        $project = factory(Project::class)
+        ->create(['started_at' => now()]);
 
         $this->assertEquals(
             $project->started_at->toDateTimeString(),
@@ -67,10 +61,8 @@ class ProjectTest extends TestCase
     /** @test */
     public function it_has_a_delivered_at_date()
     {
-        $project = Project::create([
-            'title' => 'Title',
-            'delivered_at' => now(),
-        ]);
+        $project = factory(Project::class)
+        ->create(['delivered_at' => now()]);
 
         $this->assertEquals(
             $project->delivered_at->toDateTimeString(),
@@ -81,10 +73,8 @@ class ProjectTest extends TestCase
     /** @test */
     public function it_has_a_expected_at_date()
     {
-        $project = Project::create([
-            'title' => 'Title',
-            'expected_at' => now(),
-        ]);
+        $project = factory(Project::class)
+        ->create(['expected_at' => now()]);
 
         $this->assertEquals(
             $project->expected_at->toDateTimeString(),
@@ -95,13 +85,9 @@ class ProjectTest extends TestCase
     /** @test */
     public function it_has_can_be_associated_to_a_user()
     {
-        $project = Project::create(['title' => 'Title']);
+        $project = factory(Project::class)->create(['user_id' => null]);
 
-        $user = User::create([
-            'name' => 'Test',
-            'email' => 'test@test.com',
-            'password' => Hash::make('password'),
-        ]);
+        $user = factory(User::class)->create();
 
         $this->assertNull($project->owner);
 
@@ -111,11 +97,11 @@ class ProjectTest extends TestCase
     }
 
     /** @test */
-    public function it_has_can_be_associated_to_a_status()
+    public function it_can_have_a_status()
     {
-        $project = Project::create(['title' => 'Title']);
+        $project = factory(Project::class)->create(['status_id' => null]);
 
-        $status = Status::create(['name' => 'Test']);
+        $status = factory(Status::class)->create();
 
         $this->assertNull($project->status);