+++ tests/TestCase.php
@@ -4,11 +4,11 @@ namespace Chriscreates\Projects\Tests;
 
 use Chriscreates\Projects\ProjectsServiceProvider;
 use Illuminate\Support\Facades\Schema;
-use Orchestra\Testbench\TestCase as Orchestra;
+use Orchestra\Testbench\TestCase as BaseTestCase;
 
-abstract class TestCase extends Orchestra
+abstract class TestCase extends BaseTestCase
 {
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();